In anonymous, when accessing a post: Fatal error: Call to a member function getGroups() on a non-object in htdocs/modules/imblogging/class/post.php on line 126
This is because in the beforSave event, we are accessing xoopsUser which is not declared in anonymous. But we should not get in a beforeSave event in anonymous. But we are because we ae inncrementing counter when someone view the posts. So I have added ImbloggingPost->updating_counter and when this true, the beforeSave and afterSave event are returning true without processing anything.

git-svn-id: https://subversion.assembla.com/svn/impresscmsaddons/modules/imblogging/trunk@5126 7a4d4d67-0c3a-0410-b24c-89d50c169223